Refactoring: rename commitRepository to express it is form the BD
